requirements.txt
gradio
requests
pytesseract
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
langgraph
huggingface_hub
sentence-transformers
supabase
pymupdf
wikipedia
pgvector
python-dotenv